SE AGREGA ENDPOINT PARA AÑADIR SEGUIMIENTOS Y TAMBIÉN PARA EDITAR LA INFORMACIÓN DE LOS PRODUCTORES

+++ src/main/java/com/coffe/coffeService/dto/InventarioDTO.java
@@ -2,6 +2,7 @@ package com.coffe.coffeService.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -9,7 +10,7 @@ public class InventarioDTO {
 
     private Long finca_id;
     private String producto;
-    private int cantidad;
+    private BigDecimal cantidad;
     private String unidad;
     private LocalDate fecha_registro;
 }
